# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Global Enterprise
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/dashboard:
    get:
      operationId: getDashboardStats
      tags: [admin]
      summary: Get dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Application:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        service:
          type: string
        message:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - service
        - createdAt
...
    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
      required:
        - token
        - username
    DashboardStats:
      type: object
      properties:
        totalApplications:
          type: integer
        recentApplications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        applicationsByService:
          type: array
          items:
            $ref: "#/components/schemas/ServiceCount"
        visitorCount:
          type: integer
      required:
        - totalApplications
        - recentApplications
        - applicationsByService
        - visitorCount
    ServiceCount:
      type: object
      properties:
        service:
          type: string
        count:
          type: integer
      required:
        - service
        - count
...