# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Global Enterprise
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Submit a service application
 */
export const createApplicationBodyNameMin = 2;
export const createApplicationBodyNameMax = 100;

export const createApplicationBodyPhoneMin = 10;
export const createApplicationBodyPhoneMax = 15;

export const createApplicationBodyMessageMax = 1000;

export const CreateApplicationBody = zod.object({
  name: zod
    .string()
    .min(createApplicationBodyNameMin)
    .max(createApplicationBodyNameMax),
  phone: zod
    .string()
    .min(createApplicationBodyPhoneMin)
    .max(createApplicationBodyPhoneMax),
  service: zod.enum([
    "Air Ticket Booking",
    "Train Ticket Booking",
    "International Parcel Booking",
    "PAN Card Apply",
    "Aadhaar Card Services",
    "Voter Card Apply",
  ]),
  message: zod.string().max(createApplicationBodyMessageMax).optional(),
});

/**
 * @summary List all applications (admin only)
 */
export const listApplicationsQueryLimitDefault = 50;
export const listApplicationsQueryOffsetDefault = 0;

export const ListApplicationsQueryParams = zod.object({
  service: zod.coerce.string().optional().describe("Filter by service type"),
  limit: zod.coerce.number().default(listApplicationsQueryLimitDefault),
  offset: zod.coerce.number().default(listApplicationsQueryOffsetDefault),
});

export const ListApplicationsResponse = zod.object({
  applications: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      phone: zod.string(),
      service: zod.string(),
      message: zod.string().nullish(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Export applications as CSV (admin only)
 */
export const ExportApplicationsCsvQueryParams = zod.object({
  service: zod.coerce.string().optional(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  token: zod.string(),
  username: zod.string(),
});

/**
 * @summary Get dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalApplications: zod.number(),
  recentApplications: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      phone: zod.string(),
      service: zod.string(),
      message: zod.string().nullish(),
      createdAt: zod.coerce.date(),
    }),
  ),
  applicationsByService: zod.array(
    zod.object({
      service: zod.string(),
      count: zod.number(),
    }),
  ),
  visitorCount: zod.number(),
});

/**
 * @summary Increment visitor count
 */
export const TrackVisitorResponse = zod.object({
  count: zod.number(),
});

/**
 * @summary Get total visitor count
 */
export const GetVisitorCountResponse = zod.object({
  count: zod.number(),
});